Move unpacking logic to a separate function
